# Avoid redundant copy of webhook authz cache key.