#### What type of PR is this?

/kind cleanup

#### What this PR does / why we need it:

Reduces the number of bytes allocated per call of the webhook authorizer.

Converting the []byte returned by `json.Marshal` to a string duplicates the underlying bytes (to guarantee immutability). This extra copy is specifically avoided by the implementation of `strings.Builder`. Simple benchmark included shows a per-call savings approximately equal to the length of the serialized key.

#### Which issue(s) this PR fixes:

#### Special notes for your reviewer:

Before:
```
goos: linux
goarch: amd64
pkg: k8s.io/apiserver/plugin/pkg/authorizer/webhook
cpu: Intel(R) Core(TM) i9-10850K CPU @ 3.60GHz
BenchmarkV1WebhookCache-20    	 1530652	       777.3 ns/op	     945 B/op	       6 allocs/op
PASS
ok  	k8s.io/apiserver/plugin/pkg/authorizer/webhook	1.995s
```

After:
```
goos: linux
goarch: amd64
pkg: k8s.io/apiserver/plugin/pkg/authorizer/webhook
cpu: Intel(R) Core(TM) i9-10850K CPU @ 3.60GHz
BenchmarkV1WebhookCache-20    	 1530129	       779.3 ns/op	     833 B/op	       6 allocs/op
PASS
ok  	k8s.io/apiserver/plugin/pkg/authorizer/webhook	2.003s
```

#### Does this PR introduce a user-facing change?

```release-note
NONE
```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

```docs
```